+++ Chatter/UI/Core/UIBuilder.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 using TMPro;
 
@@ -8,8 +9,8 @@ namespace ComfyLib {
   public static class UIBuilder {
     static readonly Dictionary<string, Sprite> _spriteCache = new();
 
-    static readonly Color32 _colorWhite = Color.white;
-    static readonly Color32 _colorClear = Color.clear;
+    public static readonly Color32 ColorWhite = Color.white;
+    public static readonly Color32 ColorClear = Color.clear;
 
     public static Sprite CreateRect(int width, int height, Color32 color) {
       string name = $"Rectangle-{width}w-{height}h-{color}c";
@@ -48,6 +49,82 @@ namespace ComfyLib {
       return sprite;
     }
 
+    public static Sprite CreateRoundedCornerSprite(
+        int width, int height, int radius, FilterMode filterMode = FilterMode.Bilinear) {
+      string name = $"RoundedCorner-{width}w-{height}h-{radius}r";
+
+      if (_spriteCache.TryGetValue(name, out Sprite sprite)) {
+        return sprite;
+      }
+
+      Texture2D texture =
+          new Texture2D(width, height)
+              .SetName(name)
+              .SetWrapMode(TextureWrapMode.Clamp)
+              .SetFilterMode(filterMode);
+
+      Color32[] pixels = new Color32[width * height];
+
+      for (int y = 0; y < height; y++) {
+        for (int x = 0; x < width; x++) {
+          pixels[(y * width) + x] = IsCornerPixel(x, y, width, height, radius) ? ColorClear : ColorWhite;
+        }
+      }
+
+      texture.SetPixels32(pixels);
+      texture.Apply();
+
+      int borderWidth;
+      for (borderWidth = 0; borderWidth < width; borderWidth++) {
+        if (pixels[borderWidth] == Color.white) {
+          break;
+        }
+      }
+
+      int borderHeight;
+      for (borderHeight = 0; borderHeight < height; borderHeight++) {
+        if (pixels[borderHeight * width] == Color.white) {
+          break;
+        }
+      }
+
+      sprite =
+          Sprite.Create(
+              texture,
+              new(0, 0, width, height),
+              new(0.5f, 0.5f),
+              pixelsPerUnit: 100f,
+              extrude: 0,
+              SpriteMeshType.FullRect,
+              new(borderWidth, borderHeight, borderWidth, borderHeight))
+          .SetName(name);
+
+      _spriteCache[name] = sprite;
+      return sprite;
+    }
+
+    static bool IsCornerPixel(int x, int y, int w, int h, int rad) {
+      if (rad == 0) {
+        return false;
+      }
+
+      int dx = Math.Min(x, w - x);
+      int dy = Math.Min(y, h - y);
+
+      if (dx == 0 && dy == 0) {
+        return true;
+      }
+
+      if (dx > rad || dy > rad) {
+        return false;
+      }
+
+      dx = rad - dx;
+      dy = rad - dy;
+
+      return Math.Round(Math.Sqrt(dx * dx + dy * dy)) > rad;
+    }
+
     public static Sprite CreateSuperellipse(int width, int height, float exponent) {
       string name = $"Superellipse-{width}w-{height}h-{exponent}e";
 
@@ -77,7 +154,7 @@ namespace ComfyLib {
       for (int x = 0; x < mx; x++) {
         for (int y = 0; y < my; y++) {
           float lhs = Mathf.Pow(Mathf.Abs(x / a), exponent) + Mathf.Pow(Mathf.Abs(y / b), exponent);
-          Color32 color = lhs > 1f ? _colorClear : _colorWhite;
+          Color32 color = lhs > 1f ? ColorClear : ColorWhite;
 
           int rightx = x + mx;
           int leftx = -x + mx - 1;